rm src/routes/demos.ts, move demo_routes to src/routes/+layout.ts
